+++ mz_zip_rw.c
@@ -1292,7 +1292,9 @@ int32_t mz_zip_writer_zip_cd(void *handle)
     err = mz_zip_writer_entry_open(handle, &cd_file);
     if (err == MZ_OK)
     {
-        mz_stream_copy_stream(handle, mz_zip_writer_entry_write, cd_mem_stream, NULL, (int32_t)cd_mem_length);
+        mz_stream_copy_stream(handle, mz_zip_writer_entry_write, cd_mem_stream, 
+            NULL, (int32_t)cd_mem_length);
+
         mz_stream_seek(cd_mem_stream, 0, MZ_SEEK_SET);
         mz_stream_mem_set_buffer_limit(cd_mem_stream, 0);
 
@@ -1354,7 +1356,7 @@ int32_t mz_zip_writer_entry_close(void *handle)
 #ifndef MZ_ZIP_NO_ENCRYPTION
     const uint8_t *extrafield = NULL;
     int32_t extrafield_size = 0;
-    int32_t field_length_hash = 0;
+    int16_t field_length_hash = 0;
     uint8_t sha256[MZ_HASH_SHA256_SIZE];
 
 
@@ -1447,7 +1449,9 @@ int32_t mz_zip_writer_entry_sign(void *handle, uint8_t *message, int32_t message
     if ((err == MZ_OK) && (signature != NULL))
     {
         // Write signature zip extra field
-        err = mz_zip_extrafield_write(writer->file_extra_stream, MZ_ZIP_EXTENSION_SIGN, signature_size);
+        err = mz_zip_extrafield_write(writer->file_extra_stream, MZ_ZIP_EXTENSION_SIGN,
+            (uint16_t)signature_size);
+
         if (err == MZ_OK)
         {
             if (mz_stream_write(writer->file_extra_stream, signature, signature_size) != signature_size)
@@ -1675,8 +1679,8 @@ int32_t mz_zip_writer_add_file(void *handle, const char *path, const char *filen
     return err;
 }
 
-int32_t mz_zip_writer_add_path(void *handle, const char *path, const char *root_path, uint8_t include_path, 
-    uint8_t recursive)
+int32_t mz_zip_writer_add_path(void *handle, const char *path, const char *root_path, 
+    uint8_t include_path, uint8_t recursive)
 {
     DIR *dir = NULL;
     struct dirent *entry = NULL;
